feat : 레벨 메쉬 클래스 추가, 스카이 스피어 클래스 분리
