+++ Engine/src/Engine_include.h
@@ -22,5 +22,7 @@
 #include "Engine/StaticObject.h"
 #include "Engine/Actor.h"
 #include "Engine/Level.h"
+#include "Engine/SkySphere.h"
+#include "Engine/LightMeshLevel.h"
 #include "Engine/Log.h"
 #include "Engine/EntryPoint.h"
